% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{acf_ly}
\alias{acf_ly}
\title{A Visualization Function of the ACF Estimation}
\usage{
acf_ly(ts.obj, lag.max = NULL, ci = 0.95)
}
\arguments{
\item{ts.obj}{a univariate or multivariate time series object of class "ts", "mts", "zoo" or "xts"}

\item{lag.max}{maximum lag at which to calculate the acf. Default is 10*log10(N/m) where N is the number of observations and m the number of series. Will be automatically limited to one less than the number of observations in the series.}

\item{ci}{the significant level of the estimation - a numeric value between 0 and 1, default is set for 0.95}
}
\description{
A Visualization Function of the ACF Estimation
}
\examples{
acf_ly(AirPassengers, lag.max = 60)
}
