% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Coffee_Prices}
\alias{Coffee_Prices}
\title{Coffee Prices: Robusta and Arabica}
\format{Time series data - 'mts' object}
\source{
WIKI Commodity Prices - Quandle
}
\usage{
Coffee_Prices
}
\description{
Coffee Prices: Robusta and Arabica: 1960 - 2018.
Units: Dollars per Kg
}
\examples{
ts_plot(Coffee_Prices)
}
\keyword{datasets}
