% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_functions.R
\name{create_model}
\alias{create_model}
\title{Build the Components of the \code{\link[TSstudio]{train_model}} Function}
\usage{
create_model()
}
\description{
Add, edit, or remove the components of the  \code{\link[TSstudio]{train_model}} function
}
