% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EURO_Brent}
\alias{EURO_Brent}
\title{Crude Oil Prices: Brent - Europe}
\format{Time series data - 'zoo' object}
\source{
U.S. Energy Information Administration, Crude Oil Prices: Brent - Europe [MCOILBRENTEU], retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/MCOILBRENTEU, January 8, 2018.
}
\usage{
EURO_Brent
}
\description{
Crude Oil Prices: Brent - Europe: 1987 - 2019.
Units: Dollars per Barrel
}
\examples{
ts_plot(EURO_Brent)
ts_decompose(EURO_Brent, type = "both")
}
\keyword{datasets}
