% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Michigan_CS}
\alias{Michigan_CS}
\title{University of Michigan Consumer Survey, Index of Consumer Sentiment}
\format{Time series data - 'xts' object}
\source{
University of Michigan, University of Michigan: Consumer Sentiment
}
\usage{
Michigan_CS
}
\description{
University of Michigan Consumer Survey, Index of Consumer Sentiment: 1980 - 2017.
Units: Index 1966:Q1=100
}
\examples{
ts_plot(Michigan_CS)
ts_heatmap(Michigan_CS)
}
\keyword{datasets}
