% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_plot.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plotting Forecast Object}
\usage{
plot_forecast(forecast_obj)
}
\arguments{
\item{forecast_obj}{A forecast object from the forecast package}
}
\description{
Visualization functions for forecast package forecasting objects
}
\examples{
data(USgas)
library(forecast)
fit <- forecast(ets(USgas), h = 60)
plot_forecast(fit)
}
