% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_functions.R
\name{plot_grid}
\alias{plot_grid}
\title{Visualizing Grid Search Results}
\usage{
plot_grid(grid.obj, top = NULL, highlight = 0.1, type = "parcoords",
  colors = list(showscale = TRUE, reversescale = FALSE, colorscale = "Jet"))
}
\arguments{
\item{grid.obj}{A ts_grid output object}

\item{top}{An integer, set the number of hyper-parameters combinations to visualize 
(ordered by accuracy). If set to NULL (default), will plot the top 100 combinations}

\item{highlight}{A proportion between 0 (excluding) and 1, 
set the number of hyper-parameters combinations to highlight 
(by accuracy), if the type argument is set to "parcoords"}

\item{type}{The plot type, either "3D" for 3D plot or 
"parcoords" for parallel coordinates plot}

\item{colors}{A list of plotly arguments for the color scale setting: 

showscale -  display the color scale if set to TRUE. 

reversescale - reverse the color scale if set to TRUE 

colorscale set the color scale of the plot, possible palettes are:
Greys, YlGnBu,  Greens , YlOrRd,
Bluered, RdBu, Reds, Blues, Picnic,
Rainbow, Portland, Jet, Hot, Blackbody,
Earth, Electric, Viridis, Cividis}
}
\description{
Visualizing Grid Search Results
}
