% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_functions.R
\name{plot_model}
\alias{plot_model}
\title{Plot the Models Performance on the Testing Partitions}
\usage{
plot_model(model.obj, model_ids = NULL)
}
\arguments{
\item{model.obj}{A train_model object}

\item{model_ids}{A character, defines the trained models to plot, if set to NULL (default), will plot all the models}
}
\value{
Animation of models forecast on the testing partitions compared to the actuals
}
\description{
Plot the Models Performance on the Testing Partitions
}
\details{
The plot_model provides a visualization of the models performance on the testing paritions for the train_model function output
}
