% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{reshape_ts}
\alias{reshape_ts}
\title{Transform Time Series Object to Data Frame Format}
\usage{
reshape_ts(ts.obj, type = "wide")
}
\arguments{
\item{ts.obj}{a univariate time series object of a class "ts", "zoo" or "xts" (support only series with either monthly or quarterly frequency)}

\item{type}{The reshape type - 
"wide" set the years as the columns and the cycle units (months or quarter) as the rows, or
"long" split the time object to year, cycle unit and value}
}
\description{
Transform time series object into data frame format
}
\examples{

data(USgas)
USgas_df <- reshape_ts(USgas)
}
