% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_ly.R
\name{seasonal_ly}
\alias{seasonal_ly}
\title{Seasonality Visualization of Time Series Object}
\usage{
seasonal_ly(ts.obj)
}
\arguments{
\item{ts.obj}{a univariate time series object of a class "ts", "zoo" or "xts" (support only series with either daily, monthly or quarterly frequency)}
}
\description{
Visualize time series object by it periodicity
}
\examples{
seasonal_ly(AirPassengers)
}
