% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_functions.R
\name{train_model}
\alias{train_model}
\title{Training, Testing, and Evaluating Multiple Time Series Forecasting Models}
\usage{
train_model(input, methods, train_method, horizon, xreg = NULL)
}
\arguments{
\item{input}{A univariate time series object (ts class)}

\item{methods}{A list defines a set of models to train and forecast. 
The list should include the model id, the model type, and the model arguments (please see 'details' for the structure of the argument). 
Possible models:

\code{\link[stats]{arima}} - model from the stats package 

\code{\link[forecast]{auto.arima}} - model from the forecast package

\code{\link[forecast]{ets}} - model from the forecast package}
}
\description{
Method for train test and compare multiple time series models using either one partition (i.e., sample out) 
or multipe partitions (backtesting)
}
