% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_functions.R
\name{ts_backtesting}
\alias{ts_backtesting}
\title{Evaluation Function for Forecasting Models}
\usage{
ts_backtesting(ts.obj, models = "abehntw", periods = 6, error = "MAPE",
  window_size = 3, h = 3, plot = TRUE, a.arg = NULL, b.arg = NULL,
  e.arg = NULL, h.arg = NULL, n.arg = NULL, t.arg = NULL,
  w.arg = NULL, parallel = FALSE)
}
\arguments{
\item{ts.obj}{A univariate time series object of a class "ts"}

\item{models}{String, define the type of models to use in the training function:

 'a' - auto.arima (forecast package)
 
 'b' - Bayesian Structural Time Series (bsts package)
 
 'e' - ets (forecast package) 
 
 'h' - hybrid timse series model (forecastHybrid package) 
 
 'n' - Neural Network Time Series (forecast package)
 
 't' - tbats (forecast package)
 
 'w' - Holt Winters (stats package)}

\item{periods}{The number of periods to evaluate the models (with a minimum of 2)}

\item{error}{The type of error to evaluate by - "MAPE"  (default) or "RMSE"}

\item{window_size}{An integer, the size of the backtesting window}

\item{h}{Integer, the horizon of the selected forecasting model}

\item{plot}{Logical, if TRUE desplay a plot with the backtesting progress}

\item{a.arg}{List, an optional arguments to pass to the auto.arima function}

\item{b.arg}{List, an optional arguments to pass to the bsts function, optional arguments:

linear_trend = c(TRUE, FALSE), use linear trend if set to TRUE (default)

seasonal = c(TRUE, FALSE), add seasonal component if set to TRUE (default)

niter, set the number of Markov chain Monte Carlo, default set to 1000

ping, A scalar giving the desired frequency of status messages, default set to 100 (niter/10)

family = c("gaussian", "logit", "poisson", "student"),
 The model family for the observation equation, default set to "gaussian
 
seed, An integer to use as the random seed for the underlying C++ code of the model, default set to 1234
#'}

\item{t.arg}{List, an optional arguments to pass to the tbats function}

\item{w.arg}{List, an optional arguments to pass to the Holtwinters function}

\item{parallel}{Logical, if TRUE use parallel option when applicable (auto.arima, hybridModel)}
}
\description{
Performance evaluation function for forecasting models, by training and testing the performance
of each model over a sequence of periods to identify the performance of a model over time  
(both accuracy and stability)
}
\examples{
\dontrun{
data(USgas)
USgas_backtesting <- ts_evaluate(USgas, periods = 6, window_size = 24, h = 60, error = "RMSE")

# Selecting a specific models (auto.arima, ets and nnetar)
USgas_backtesting <- ts_evaluate(USgas, models = "aen", periods = 6, window_size = 24, h = 60)

# Using some of the models arguments 
 USgas_backtesting <- ts_evaluate(USgas, models = "aen", periods = 6, window_size = 24, h = 60, error = "MAPE"
 a.arg = list(stepwise = FALSE), n.arg = list(P =2 , p = 12, repeats = 50))

# Retrieve the models leaderboard
USgas_backtesting$leaderboard

# Retrieve the best forecast results
USgas_backtesting$leadForecast$mean

# Retrieve the final forecast of the ets model
USgas_backtesting$Forecast_Final$ets$mean

# Retrieve the ets forecast during the first period of testing
USgas_backtesting$period_1$ets$forecast$mean

# Get a summary of the models RMSE results for each testing period
USgas_backtesting$RMSE_score

# Get the final plot of the models performance and the selected forecasting model
USgas_backtesting$summary_plot


}
}
