% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_functions.R
\name{ts_evaluate}
\alias{ts_evaluate}
\title{Evaluation Function for Forecasting Models}
\usage{
ts_evaluate(ts.obj, models = "abehntw", periods = 6, error = "MAPE",
  h = 3)
}
\arguments{
\item{ts.obj}{A univariate time series object of a class "ts"}

\item{models}{String, define the type of models to use in the training function:
'a' - auto.arima (forecast package)
'b' - Bayesian Structural Time Series (bsts package)
'e' - ETS (forecast package) 
'h' - Hybrid (forecastHybrid package) 
'n' - Neural Network Time Series (forecast package)
't' - TBATS (forecast package)
'w' - Holt Winters (stats package)}

\item{periods}{The number of periods to evaluate the models (with a minimum of 2)}

\item{error}{The type of error to evaluate by - "MAPE"  (default) or "RMSE"}

\item{h}{Integer, the horizon of the forecast}
}
\description{
Performance evaluation function for forecasting models, by training and testing the performance
of each model over a sequence of periods to identify the performance of a model over time  
(both accuracy and stability)
}
\examples{
data(USgas)
ts_evaluate(USgas, periods = 12, h = 12)
}
