% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{ts_info}
\alias{ts_info}
\title{Get the Time Series Information}
\usage{
ts_info(ts.obj)
}
\arguments{
\item{ts.obj}{A time series object of a class "ts", "mts", "xts", or "zoo"}
}
\value{
Text
}
\description{
Returning the time series object main characteristics
}
\examples{

# ts object
data("USgas")
ts_info(USgas)

# mts object
data("Coffee_Prices")
get_info(Coffee_Prices)

# xts object
data("Michigan_CS")
get_info(Michigan_CS)
}
