% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_seasonal.R
\name{ts_polar}
\alias{ts_polar}
\title{Polor Plot for Time Series Object}
\usage{
ts_polar(ts.obj, title = NULL, width = 600, height = 600, left = 25,
  right = 25, top = 25, bottom = 25)
}
\arguments{
\item{ts.obj}{A univariate time series object of a class "ts", "zoo" or "xts" (support only series with either monthly or quarterly frequency)}

\item{title}{Add a title for the plot, default set to NULL}

\item{width}{The widht of the plot in pixels, default set to 600}

\item{height}{The height of the plot pixels, default set to 600}

\item{left}{Set the left margin of the plot in pixels, default set to 25}

\item{right}{Set the right margin of the plot in pixels, default set to 25}

\item{top}{Set the top margin of the plot in pixels, default set to 25}

\item{bottom}{Set the bottom margin of the plot in pixels, default set to 25}
}
\description{
Polor plot for time series object (ts, zoo, xts), currently support only monthly and quarterly frequency
}
\examples{
data(USgas)
ts_polar(USgas)
}
