% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_seasonal.R
\name{ts_quantile}
\alias{ts_quantile}
\title{Quantile Plot for Time Series}
\usage{
ts_quantile(ts.obj, upper = 0.75, lower = 0.25, period = NULL, n = 1)
}
\arguments{
\item{ts.obj}{A univariate time series object of a class "zoo", "xts", or data frame family ("data.frame", "data.table", "tbl")}

\item{upper}{A numeric value between 0 and 1 (excluding 0, and greater than the "lower" argument) set the upper bound of the quantile plot 
(using the "probs" argument of the \code{\link[stats]{quantile}} function). By default set to 0.75}

\item{lower}{A numeric value between 0 and 1 (excluding 1, and lower than the "upper" argument) set the upper bound of the quantile plot 
(using the "probs" argument of the \code{\link[stats]{quantile}} function). By default set to 0.25}

\item{period}{A character, set the period level of the data for the quantile calculation and plot representation. 
Must be one level above the input frequency (e.g., an hourly data can represent by daily, weekdays, monthly, quarterly and yearly). 
Possible options c("daily", "weekdays", "monthly", "quarterly", "yearly")}

\item{n}{An integer, set the number of plots rows to display (by setting the nrows argument in the \code{\link[plotly]{subplot}} function), must be an integer between 1 and the frequency of the period argument.}

\item{type}{A character, set the display mode of the plot to either single plot for all periods ("single") or plot for each period ("multiple", default option)}
}
\description{
A quantile plot of time series data, allows the user to display a quantile plot of a series by a subset period
}
\examples{


}
