% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_seasonal.R
\name{ts_seasonal}
\alias{ts_seasonal}
\title{Seasonality Visualization of Time Series Object}
\usage{
ts_seasonal(ts.obj, type = "normal", Ygrid = FALSE, Xgrid = FALSE,
  title = NULL, last = NULL)
}
\arguments{
\item{ts.obj}{A univariate time series object of a class "ts", "zoo", or "xts" 
(support only series with either monthly or quarterly frequency). 
Also, this function supports data frame objects as long as there is at least one "Date" and "numeric" objects 
(if there are more, by defualt will use the first of each)}

\item{type}{The type of the seasonal plot - 
"normal" to split the series by full cycle units, or
"cycle" to split by cycle units, or
"box" for box-plot by cycle units, or
"all" for all the three plots together}

\item{Ygrid}{Logic,show the Y axis grid if set to TRUE}

\item{Xgrid}{Logic,show the X axis grid if set to TRUE}

\item{title}{Plot title - Character object}

\item{last}{Subset the data to the last number of observations}
}
\description{
Visualize time series object by it periodicity, currently support only monthly and quarterly frequency
}
\examples{
data(USgas)
ts_seasonal(USgas)

# Seasonal box plot
ts_seasonal(USgas, type = "box") 
}
