% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{ts_split}
\alias{ts_split}
\title{Split Time Series Object for Training and Testing Partitions}
\usage{
ts_split(ts.obj, sample.out = NULL)
}
\arguments{
\item{ts.obj}{A univariate time series object of a class "ts"}

\item{sample.out}{An integer, set the number of periods of the testing or sample out partition, defualt set for 
30 percent of the lenght of the series}
}
\description{
Split a time series object into training and testing partitions
}
\examples{

## Split the USgas dataset into training and testing partitions

## Set the last 12 months as a testing partition 

## and the rest as a training partition

data(USgas, package = "TSstudio")

split_USgas <- ts_split(ts.obj = USgas, sample.out = 12)

training <- split_USgas$train
testing <- split_USgas$test

length(USgas)

length(training)
length(testing)
}
