% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{ts_sum}
\alias{ts_sum}
\title{Summation of Multiple Time Series Object}
\usage{
ts_sum(mts.obj)
}
\arguments{
\item{mts.obj}{A multivariate time series object of a class "mts"}
}
\description{
A row sum function for multiple time series object ("mts"), return the
the summation of the "mts" object as a "ts" object
}
\examples{

x <- matrix(c(1:100, 1:100, 1:100), ncol = 3)
mts.obj <- ts(x, start = c(2000, 1), frequency = 12)
ts_total <- ts_sum(mts.obj)

}
