% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_seasonal.R
\name{ts_surface}
\alias{ts_surface}
\title{3D Surface Plot for Time Series}
\usage{
ts_surface(ts.obj)
}
\arguments{
\item{ts.obj}{a univariate time series object of a class "ts", "zoo" or "xts" (support only series with either monthly or quarterly frequency)}
}
\description{
3D surface plot for time series object by it periodicity (currently support only monthly and quarterly frequency)
}
\examples{
ts_surface(USgas) 
}
