% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USgas}
\alias{USgas}
\title{US monthly natural gas consumption}
\format{Time series data - 'ts' object}
\source{
U.S. Bureau of Transportation Statistics, Natural Gas Consumption [NATURALGAS], retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/NATURALGAS, January 7, 2018.
}
\usage{
USgas
}
\description{
US monthly natural gas consumption: 2000 - 2017.
Units: Billion Cubic Feet
}
\examples{
ts_plot(USgas)
ts_seasonal(USgas, type = "all")
}
\keyword{datasets}
