% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USVSales}
\alias{USVSales}
\title{US Monthly Total Vehicle Sales}
\format{Time series data - 'ts' object}
\source{
U.S. Bureau of Economic Analysis, Total Vehicle Sales [TOTALNSA], retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/TOTALNSA, January 7, 2018.
}
\usage{
USVSales
}
\description{
US monthly total vehicle sales: 1976 - 2017.
Units: Thousands of units
}
\examples{
ts_plot(USVSales)
seasonal_ly(USVSales)
}
\keyword{datasets}
