% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_functions.R
\name{create_model}
\alias{create_model}
\title{Build the \code{\link[TSstudio]{train_model}} Function Components}
\usage{
create_model()
}
\description{
Build the \code{\link[TSstudio]{train_model}} Function Components
}
