% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_plot.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plotting Forecast Object}
\usage{
plot_forecast(forecast_obj, title = NULL, Xtitle = NULL,
  Ytitle = NULL, color = NULL, width = 2)
}
\arguments{
\item{forecast_obj}{A forecast object from the forecast, forecastHybrid, or bsts packages}

\item{title}{A character, a plot title, optional}

\item{Xtitle}{Set the X axis title, default set to NULL}

\item{Ytitle}{Set the Y axis title, default set to NULL}

\item{color}{A character, the plot, support both name and expression}

\item{width}{An Integer, define the plot width, default is set to 2}
}
\description{
Visualization functions for forecast package forecasting objects
}
\examples{
data(USgas)
library(forecast)
fit <- ets(USgas)
fc<- forecast(fit, h = 60)
plot_forecast(fc)
}
