% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_functions.R
\name{ts_grid}
\alias{ts_grid}
\title{Evaluation Function for Forecasting Models}
\usage{
ts_grid(ts.obj, model, periods, window_length = NULL, window_space,
  window_test, hyper_params, search_criteria, parallel = TRUE,
  n.cores = "auto")
}
\arguments{
\item{ts.obj}{A univariate time series object of a class "ts"}

\item{model}{A string, defines the model}
}
\description{
Evaluation Function for Forecasting Models
}
