% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_seasonal.R
\name{ts_heatmap}
\alias{ts_heatmap}
\title{Heatmap Plot for Time Series}
\usage{
ts_heatmap(ts.obj, last = NULL)
}
\arguments{
\item{ts.obj}{A univariate time series object of a class "ts", "zoo" or "xts" (support only series with either monthly or quarterly frequency)}

\item{last}{Subset the data to the last number of observations}
}
\description{
Heatmap plot for time series object by it periodicity (currently support only monthly and quarterly frequency)
}
\examples{
data(USgas)
ts_heatmap(USgas) 
}
