% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{ts_lags}
\alias{ts_lags}
\title{Time Series Lag Visualization}
\usage{
ts_lags(ts.obj, lag.max = 12, Xtitle = FALSE, Ytitle = TRUE,
  margin = 0.02, Xshare = TRUE, Yshare = TRUE, n_row = 3)
}
\arguments{
\item{ts.obj}{A univariate time series object of a class "ts", "zoo" or "xts" (support only series with either monthly or quarterly frequency)}

\item{lag.max}{An integer, number of lags to plot}

\item{Xtitle}{Plotly parameter, should x-axis titles be retained?}

\item{Ytitle}{Plotly parameter, should y-axis titles be retained?}

\item{margin}{Plotly parameter, either a single value or four values (all between 0 and 1). 
If four values are provided, the first is used as the left margin, 
the second is used as the right margin, the third is used as the top margin, 
and the fourth is used as the bottom margin. 
If a single value is provided, it will be used as all four margins.}

\item{Xshare}{Plotly parameter, should the x-axis be shared amongst the subplots?}

\item{Yshare}{Plotly parameter, should the y-axis be shared amongst the subplots?}

\item{n_row}{An integer, define the number of plots per row}
}
\description{
Visualization of series with its lags, 
can be used to identify a correlation between the series and it lags
}
\examples{
# Seasonal box plot
ts_lags(AirPassengers) 
}
