% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_seasonal.R
\name{ts_ma}
\alias{ts_ma}
\title{Moving Average Time Series Data}
\usage{
ts_ma(ts.obj, k = c(3, 6, 9), k_left = NULL, k_right = NULL,
  double = NULL, plot = TRUE, multiple = FALSE, title = NULL,
  Xtitle = NULL, Ytitle = NULL)
}
\arguments{
\item{ts.obj}{a univariate time series object of a class "ts", "zoo" or "xts" (support only series with either monthly or quarterly frequency)}

\item{k}{A single or multiple integers (by default using 3, 6, and 9 as inputs), 
define a two-sides moving averages by setting the number of past and future to use 
in each moving average window along with current observation.}

\item{k_left}{A single integer (optional argument, default set to NULL), can be used, 
along with the k_right argument, an unbalanced moving average. 
The k_left defines the number of lags to includes in the moving average.}

\item{k_right}{A single integer (optional argument, default set to NULL), can be used, 
along with the k_left argument, to set an unbalanced moving average. 
The k_right defines the number of negative lags to includes in the moving average.}

\item{double}{A single integer, an optional argument. If not NULL (by default), will apply a second moving average process on the initial moving average output}

\item{plot}{A boolean, if TRUE will plot the results}

\item{multiple}{A boolean, if TRUE (and k > 1) will create multiple plots, one for each moving average degree. By default is set to FALSE}

\item{title}{A character, if not NULL (by default), will use the input as the plot title}

\item{Xtitle}{A character, if not NULL (by default), will use the input as the plot x - axis title}

\item{Ytitle}{A character, if not NULL (by default), will use the input as the plot y - axis title}
}
\description{
Calculate the moving average (and double moving average) for time series data
}
\examples{

# Using moving average to smooth the USVsales dataset (US Monthly Total Vehicle Sales)
# Applying 4 degrees of moving average   
USVS_MA <- ts_ma(ts.obj = USVSales, k = c(3, 6, 9, 12))

# Display each moving average (by degree) on a separate plot and adding a title for the plot
USVS_MA <- ts_ma(ts.obj = USVSales, k = c(3, 6, 9, 12), 
                 multiple = TRUE, 
                 title = "Moving Average for the US Vehical Sales Series")

# Plot only the moving average with k = 12
ts_plot(USVS_MA$ma_12, title = "Moving Average, k = 12")

# Using moving average to smooth the USgas dataset (US Monthly Natural Gas Consuption)
# Adding double moving average with degree of 6 to smooth the strong seasonality
USgas_MA <- ts_ma(ts.obj = USgas, k = c(6, 9), double = 6)

# Extract the plot
USgas_MA$plot
}
