% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_ly.R
\name{ts_polar}
\alias{ts_polar}
\title{Polor Plot for Time Series Object}
\usage{
ts_polar(ts.obj, title = NULL, width = 600, height = 600, left = 25,
  right = 25, top = 25, bottom = 25)
}
\arguments{
\item{ts.obj}{a univariate time series object of a class "ts", "zoo" or "xts" (support only series with either monthly or quarterly frequency)}

\item{width}{the widht of the plot in pixels, default set to 600}

\item{height}{the height of the plot pixels, default set to 600}

\item{type}{The type of the seasonal plot - 
"normal" to split the series by full cycle units, or
"cycle" to split by cycle units, or
"box" for box-plot by cycle units, or
"polar"  for polar plot}
}
\description{
Polor plot for time series object (ts, zoo, xts), currently support only monthly and quarterly frequency
}
\examples{
data(USgas)
ts_polar(USgas)
}
