% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USUnRate}
\alias{USUnRate}
\title{US Monthly Civilian Unemployment Rate}
\format{Time series data - 'ts' object}
\source{
U.S. Bureau of Labor Statistics, Civilian Unemployment Rate [UNRATENSA], retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/UNRATENSA, January 6, 2018.
}
\usage{
USUnRate
}
\description{
US monthly civilian unemployment rate: 1948 - 2019.
Units: Percent
}
\examples{
ts_plot(USUnRate)
ts_seasonal(USUnRate)
}
\keyword{datasets}
