% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{xts_to_ts}
\alias{xts_to_ts}
\title{Converting 'xts' object to 'ts' object}
\usage{
xts_to_ts(xts.obj)
}
\arguments{
\item{xts.obj}{a univariate 'xts' object}
}
\description{
Converting 'xts' object to 'ts' object
}
\examples{

data("Michigan_CS", package = "TSstudio")
class(Michigan_CS)
ts_plot(Michigan_CS)
Michigan_CS_ts <- xts_to_ts(Michigan_CS)
ts_plot(Michigan_CS_ts)
}
