% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{zoo_to_ts}
\alias{zoo_to_ts}
\title{Converting 'zoo' object to 'ts' object}
\usage{
zoo_to_ts(zoo.obj)
}
\arguments{
\item{zoo.obj}{a univariate 'zoo' object}
}
\description{
Converting 'zoo' object to 'ts' object
}
\examples{

data("EURO_Brent", package = "TSstudio")
class(EURO_Brent)
ts_plot(EURO_Brent)
EURO_Brent_ts <- zoo_to_ts(EURO_Brent)
class(EURO_Brent_ts)
ts_plot(EURO_Brent_ts)
}
